SBI-19: Extract MutableClock to testFixtures for reuse
Move MutableClock from RpcUrlManagerTest inner class to
src/testFixtures/testutil/ so it can be reused by worker
engine and other time-dependent tests.
